requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
paho-mqtt==1.6.1
sqlalchemy==1.4.52
pydantic==1.10.7
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
PyJWT==2.7.0
pymysql==1.0.3
alembic==1.11.1
redis==4.5.5
httpx==0.24.1
python-multipart==0.0.6
